pyproject.toml
[project]
name = "model-benchmark-agent"
version = "0.1.0"
description = "Model benchmarking and optimization agent with Gradio frontend"
authors = [
    {name = "Nishant", email = "bazigarofai@gmail.com"}
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "accelerate>=0.20.0",
    "optimum-quanto>=0.2.0",
    "datasets>=2.12.0",
    "gradio>=4.0.0",
    "psutil>=5.9.0",
    "numpy>=1.24.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
]
requires-python = ">=3.10"

# For Future Release
#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]